requirements.txt
flaskflask
psycopg2-binary
matplotlib